+++ tests/testthat/test-mice.R
@@ -1,22 +1,22 @@
-context('mice')
+# context('mice')
 
-# load
-library(modelsummary)
-library(mice)
+# # load
+# library(modelsummary)
+# library(mice)
 
-# impute
-set.seed(10393983)
-dat <- mtcars
-dat$wt[sample(1:nrow(dat), 4)] <- NA
-dat$hp[sample(1:nrow(dat), 3)] <- NA
-dat <- mice(dat, m = 5, printFlag = FALSE, seed = 10)
+# # impute
+# set.seed(10393983)
+# dat <- mtcars
+# dat$wt[sample(1:nrow(dat), 4)] <- NA
+# dat$hp[sample(1:nrow(dat), 3)] <- NA
+# dat <- mice(dat, m = 5, printFlag = FALSE, seed = 10)
 
-# fit
-f <- am ~ wt + hp
-mod <- list()
-mod$OLS <- with(dat, lm(am ~ wt + hp + vs))
-mod$Logit <- with(dat, glm(am ~ wt + vs, family = binomial()))
-mod <- lapply(mod, mice::pool)
+# # fit
+# f <- am ~ wt + hp
+# mod <- list()
+# mod$OLS <- with(dat, lm(am ~ wt + hp + vs))
+# mod$Logit <- with(dat, glm(am ~ wt + vs, family = binomial()))
+# mod <- lapply(mod, mice::pool)
 
 # # test
 # test_that("ols and logit", {