+++ NEWS.md
@@ -1,12 +1,13 @@
 # modelsummary 0.6.3.9000
 
-* glue format + deprecate statistic_vertical
+* glue format for `estimate` and `statistic`
 * support for easystats packages for model info extraction: performance, parameters
+* deprecate statistic_vertical
+* deprecate extract_models. Use modelsummary(output="dataframe") instead.
 * modelplot pushes ... through to extract_models
 * datasummary_skim(type="dataset")
 * gof_map omits by default
 * statistic_override does not require a list
-* extract_models is deprecated. Use modelsummary(output="dataframe") instead.
 * datasummary_balance uses row percentages
 * statistic_override accepts a single model
 * N function for well formatted N in datasummary